+++ tests/test_transaction.py
@@ -2,6 +2,7 @@ import unittest
 from unittest import mock
 from unittest.mock import call
 
+from flumine.clients import ExchangeType
 from flumine.execution.transaction import Transaction, OrderPackageType
 from flumine.exceptions import ControlError, OrderError
 
@@ -10,7 +11,7 @@ class TransactionTest(unittest.TestCase):
     def setUp(self) -> None:
         mock_blotter = {}
         self.mock_market = mock.Mock(blotter=mock_blotter)
-        self.mock_client = mock.Mock(trading_controls=[])
+        self.mock_client = mock.Mock(trading_controls=[], EXCHANGE=ExchangeType.BETFAIR)
         self.transaction = Transaction(self.mock_market, 1, False, self.mock_client)
 
     def test_init(self):
@@ -179,7 +180,7 @@ class TransactionTest(unittest.TestCase):
         return_value=True,
     )
     def test_update_order(self, mock__validate_controls):
-        mock_order = mock.Mock(client=self.mock_client)
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=ExchangeType.BETFAIR)
         self.assertTrue(self.transaction.update_order(mock_order, "PERSIST"))
         mock_order.update.assert_called_with("PERSIST")
         mock__validate_controls.assert_called_with(mock_order, OrderPackageType.UPDATE)
@@ -187,7 +188,7 @@ class TransactionTest(unittest.TestCase):
         self.assertTrue(self.transaction._pending_orders)
 
     def test_update_order_incorrect_client(self):
-        mock_order = mock.Mock(client=123)
+        mock_order = mock.Mock(client=123, EXCHANGE=ExchangeType.BETFAIR)
         with self.assertRaises(OrderError):
             self.transaction.update_order(mock_order, "PERSIST")
 
@@ -196,18 +197,23 @@ class TransactionTest(unittest.TestCase):
         return_value=False,
     )
     def test_update_order_violation(self, mock__validate_controls):
-        mock_order = mock.Mock(client=self.mock_client)
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=ExchangeType.BETFAIR)
         self.assertFalse(self.transaction.update_order(mock_order, "test"))
         mock__validate_controls.assert_called_with(mock_order, OrderPackageType.UPDATE)
         self.transaction._pending_update = []
         self.assertFalse(self.transaction._pending_orders)
+
+    def test_update_order_unknown_exchange(self):
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=123)
+        with self.assertRaises(OrderError):
+            self.transaction.update_order(mock_order, "PERSIST", force=True)
 
     @mock.patch(
         "flumine.execution.transaction.Transaction._validate_controls",
         return_value=False,
     )
     def test_force_update_order(self, mock__validate_controls):
-        mock_order = mock.Mock(client=self.mock_client)
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=ExchangeType.BETFAIR)
         self.assertTrue(
             self.transaction.update_order(mock_order, "PERSIST", force=True)
         )
@@ -220,8 +226,31 @@ class TransactionTest(unittest.TestCase):
         "flumine.execution.transaction.Transaction._validate_controls",
         return_value=True,
     )
+    def test_update_order_betdaq(self, mock__validate_controls):
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=ExchangeType.BETDAQ)
+        self.assertTrue(
+            self.transaction.update_order(
+                mock_order,
+                size_delta=-1,
+                new_price=2,
+                expected_selection_reset_count=3,
+                expected_withdrawal_sequence_number=4,
+                cancel_on_in_running=False,
+                cancel_if_selection_reset=False,
+                set_to_be_sp_if_unmatched=True,
+            )
+        )
+        mock_order.update.assert_called_with(-1, 2, 3, 4, False, False, True)
+        mock__validate_controls.assert_called_with(mock_order, OrderPackageType.UPDATE)
+        self.transaction._pending_update = [(mock_order, None)]
+        self.assertTrue(self.transaction._pending_orders)
+
+    @mock.patch(
+        "flumine.execution.transaction.Transaction._validate_controls",
+        return_value=True,
+    )
     def test_replace_order(self, mock__validate_controls):
-        mock_order = mock.Mock(client=self.mock_client)
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=ExchangeType.BETFAIR)
         self.assertTrue(self.transaction.replace_order(mock_order, 1.01, 321))
         mock_order.replace.assert_called_with(1.01)
         mock__validate_controls.assert_called_with(mock_order, OrderPackageType.REPLACE)
@@ -229,7 +258,12 @@ class TransactionTest(unittest.TestCase):
         self.assertTrue(self.transaction._pending_orders)
 
     def test_replace_order_incorrect_client(self):
-        mock_order = mock.Mock(client=123)
+        mock_order = mock.Mock(client=123, EXCHANGE=ExchangeType.BETFAIR)
+        with self.assertRaises(OrderError):
+            self.transaction.replace_order(mock_order, 1.01, 321)
+
+    def test_replace_order_incorrect_exchange(self):
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=123)
         with self.assertRaises(OrderError):
             self.transaction.replace_order(mock_order, 1.01, 321)
 
@@ -238,7 +272,7 @@ class TransactionTest(unittest.TestCase):
         return_value=False,
     )
     def test_replace_order_violation(self, mock__validate_controls):
-        mock_order = mock.Mock(client=self.mock_client)
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=ExchangeType.BETFAIR)
         self.assertFalse(self.transaction.replace_order(mock_order, 2.02))
         mock__validate_controls.assert_called_with(mock_order, OrderPackageType.REPLACE)
         self.transaction._pending_replace = []
@@ -249,7 +283,7 @@ class TransactionTest(unittest.TestCase):
         return_value=False,
     )
     def test_force_replace_order(self, mock__validate_controls):
-        mock_order = mock.Mock(client=self.mock_client)
+        mock_order = mock.Mock(client=self.mock_client, EXCHANGE=ExchangeType.BETFAIR)
         self.assertTrue(
             self.transaction.replace_order(mock_order, 1.01, 321, force=True)
         )
@@ -347,17 +381,31 @@ class TransactionTest(unittest.TestCase):
         mock_client_control.assert_not_called()
 
     @mock.patch("flumine.execution.transaction.BetfairOrderPackage")
-    def test__create_order_package(self, mock_betfair_order_package):
+    def test__create_order_package_betfair(self, mock_betfair_order_package):
         mock_betfair_order_package.order_limit.return_value = 2
+        mock_order_one = mock.Mock(id=1, exchange="BETFAIR")
+        mock_order_two = mock.Mock(id=2, exchange="BETFAIR")
+        mock_order_three = mock.Mock(id=3, exchange="BETFAIR")
+        mock_order_four = mock.Mock(id=4, exchange="BETFAIR")
+        mock_order_five = mock.Mock(id=5, exchange="BETFAIR")
         packages = self.transaction._create_order_package(
-            [(1, None), (2, None), (3, 123), (4, 123), (5, 123)], OrderPackageType.PLACE
+            [
+                (mock_order_one, None),
+                (mock_order_two, None),
+                (mock_order_three, 123),
+                (mock_order_four, 123),
+                (mock_order_five, 123),
+            ],
+            OrderPackageType.PLACE,
         )
+        self.assertEqual(len(packages), 3)
         mock_betfair_order_package.assert_has_calls(
             [
+                call.order_limit(OrderPackageType.PLACE),
                 call(
                     client=self.transaction._client,
                     market_id=self.transaction.market.market_id,
-                    orders=[1, 2],
+                    orders=[mock_order_one, mock_order_two],
                     package_type=OrderPackageType.PLACE,
                     bet_delay=self.transaction.market.market_book.bet_delay,
                     market_version=None,
@@ -366,7 +414,7 @@ class TransactionTest(unittest.TestCase):
                 call(
                     client=self.transaction._client,
                     market_id=self.transaction.market.market_id,
-                    orders=[3, 4],
+                    orders=[mock_order_three, mock_order_four],
                     package_type=OrderPackageType.PLACE,
                     bet_delay=self.transaction.market.market_book.bet_delay,
                     market_version=123,
@@ -375,7 +423,7 @@ class TransactionTest(unittest.TestCase):
                 call(
                     client=self.transaction._client,
                     market_id=self.transaction.market.market_id,
-                    orders=[5],
+                    orders=[mock_order_five],
                     package_type=OrderPackageType.PLACE,
                     bet_delay=self.transaction.market.market_book.bet_delay,
                     market_version=123,
@@ -383,9 +431,85 @@ class TransactionTest(unittest.TestCase):
                 ),
             ]
         )
-        self.assertEqual(len(packages), 3)
-        mock_betfair_order_package.order_limit.assert_called_with(
-            OrderPackageType.PLACE
+
+    @mock.patch("flumine.execution.transaction.BetfairOrderPackage")
+    def test__create_order_package_simulated(self, mock_betfair_order_package):
+        self.transaction._client.EXCHANGE = ExchangeType.SIMULATED
+        mock_betfair_order_package.order_limit.return_value = 2
+        mock_order_one = mock.Mock(id=1, exchange="SIMULATED")
+        mock_order_two = mock.Mock(id=2, exchange="SIMULATED")
+        mock_order_three = mock.Mock(id=3, exchange="SIMULATED")
+        packages = self.transaction._create_order_package(
+            [
+                (mock_order_one, None),
+                (mock_order_two, None),
+                (mock_order_three, None),
+            ],
+            OrderPackageType.PLACE,
+        )
+        self.assertEqual(len(packages), 2)
+        mock_betfair_order_package.assert_has_calls(
+            [
+                call.order_limit(OrderPackageType.PLACE),
+                call(
+                    client=self.transaction._client,
+                    market_id=self.transaction.market.market_id,
+                    orders=[mock_order_one, mock_order_two],
+                    package_type=OrderPackageType.PLACE,
+                    bet_delay=self.transaction.market.market_book.bet_delay,
+                    market_version=None,
+                    async_=False,
+                ),
+                call(
+                    client=self.transaction._client,
+                    market_id=self.transaction.market.market_id,
+                    orders=[mock_order_three],
+                    package_type=OrderPackageType.PLACE,
+                    bet_delay=self.transaction.market.market_book.bet_delay,
+                    market_version=None,
+                    async_=False,
+                ),
+            ]
+        )
+
+    @mock.patch("flumine.execution.transaction.BetdaqOrderPackage")
+    def test__create_order_package_betdaq(self, mock_betdaq_order_package):
+        self.transaction._client.EXCHANGE = ExchangeType.BETDAQ
+        mock_betdaq_order_package.order_limit.return_value = 2
+        mock_order_one = mock.Mock(id=1, exchange="BETDAQ")
+        mock_order_two = mock.Mock(id=2, exchange="BETDAQ")
+        mock_order_three = mock.Mock(id=3, exchange="BETDAQ")
+        packages = self.transaction._create_order_package(
+            [
+                (mock_order_one, None),
+                (mock_order_two, None),
+                (mock_order_three, None),
+            ],
+            OrderPackageType.PLACE,
+        )
+        self.assertEqual(len(packages), 2)
+        mock_betdaq_order_package.assert_has_calls(
+            [
+                call.order_limit(OrderPackageType.PLACE),
+                call(
+                    client=self.transaction._client,
+                    market_id=self.transaction.market.market_id,
+                    orders=[mock_order_one, mock_order_two],
+                    package_type=OrderPackageType.PLACE,
+                    bet_delay=self.transaction.market.market_book.bet_delay,
+                    market_version=None,
+                    async_=False,
+                ),
+                call(
+                    client=self.transaction._client,
+                    market_id=self.transaction.market.market_id,
+                    orders=[mock_order_three],
+                    package_type=OrderPackageType.PLACE,
+                    bet_delay=self.transaction.market.market_book.bet_delay,
+                    market_version=None,
+                    async_=False,
+                ),
+            ]
         )
 
     @mock.patch("flumine.execution.transaction.Transaction.execute")