+++ HISTORY.rst
@@ -3,6 +3,17 @@
 Release History
 ---------------
 
+2.7.0 (2025-02-19)
++++++++++++++++++++
+
+**Improvements**
+
+- BETDAQ Retail client integration (account/execution/order only)
+
+**Libraries**
+
+- betdaq-retail 0.1.0 added as dependency
+
 2.6.10 (2025-02-07)
 +++++++++++++++++++